v2

+++ database/migrations/2022_09_11_223298_create_types_table.php
@@ -1,8 +1,8 @@
 <?php
 
-use Illuminate\Support\Facades\Schema;
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
@@ -13,31 +13,32 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('types', function (Blueprint $table) {
-            $table->id();
+        if (! Schema::hasTable('types')) {
+            Schema::create('types', function (Blueprint $table) {
+                $table->id();
 
-            $table->foreignId('parent_id')->nullable()->references('id')->on('types')->onDelete('cascade');
+                $table->foreignId('parent_id')->nullable()->references('id')->on('types')->onDelete('cascade');
 
+                //Morph
+                $table->string('model_type')->nullable();
+                $table->unsignedBigInteger('model_id')->nullable();
 
-            //Morph
-            $table->string('model_type')->nullable();
-            $table->unsignedBigInteger('model_id')->nullable();
+                //Types
+                $table->string('for')->default('posts')->nullable();
+                $table->string('type')->default('category')->nullable();
 
-            //Types
-            $table->string('for')->default('posts')->nullable();
-            $table->string('type')->default('category')->nullable();
+                $table->string('name');
+                $table->string('key')->unique()->index();
+                $table->text('description')->nullable();
 
-            $table->string('name');
-            $table->string('key')->unique()->index();
-            $table->text('description')->nullable();
+                //Icon & Color
+                $table->string('color')->nullable();
+                $table->string('icon')->nullable();
 
-            //Icon & Color
-            $table->string('color')->nullable();
-            $table->string('icon')->nullable();
-
-            $table->boolean('is_activated')->default(true)->nullable();
-            $table->timestamps();
-        });
+                $table->boolean('is_activated')->default(true)->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
