+++ database/migrations/2023_02_13_143941_create_typables_table.php
@@ -1,8 +1,8 @@
 <?php
 
-use Illuminate\Support\Facades\Schema;
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
@@ -13,12 +13,14 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('typables', function (Blueprint $table) {
-            $table->foreignId("type_id")->references('id')->on('types')->onDelete('cascade');
+        if (! Schema::hasTable('typables')) {
+            Schema::create('typables', function (Blueprint $table) {
+                $table->foreignId('type_id')->references('id')->on('types')->onDelete('cascade');
 
-            $table->unsignedBigInteger("typables_id");
-            $table->string("typables_type");
-        });
+                $table->unsignedBigInteger('typables_id');
+                $table->string('typables_type');
+            });
+        }
     }
 
     /**
